perf: convert file tree git status once
